Initial work on choosable logger levels
Right now, the appropriate NetworkTables values are being sent to the
dashboard but the dashboard mysteriously no longer displays the actual
radio button chooser.

Using SmartDashboard SFX, the raw options can be seen in an ArrayView(?)
but I can't figure out how to show them on the normal dashboard.

Potential lead: SendableChooser reports its display type as "String
Chooser", we may have to do that explicitly in LoggerChooser even though
it extends that method and should report the same.