Added PID and IMU turning, and removed DriveTicks. Revert revert
This is a revert of a revert so that my slight Git misconfiguration is
no loger a problem, but so that we also get the changes from a previous
commit. We now have everything in place for PID and IMU based turning including
a command and multiple methods in drivetrain for said command. PID
values still need to be tuned.
DriveTicks methods in Drivetrain and the DriveTicksCommand has been
removed, which I am now rethinking. This was removed because it had
become obsolete as a result of work on Position mode based movement with
encoders, along with the fact that this functionality didn't really work
well and was only intended as a test.
Changes are also showing up in a number of other files that I didn't
manually touch, but got changed by Eclipse's automatic code formatiting
which I applied to the entire project.