Implement motion replay
This is a port of
https://github.com/Spartronics4915/2017-Demo-ActionReplay to the real
codebase.  One addition is that the program is stored and loaded from
SmartDashboard in two double[]s named Drivetrain_Replay_Forward and
Drivetrain_Replay_Rotation.  These need to be completely empty strings
to save a recording, otherwise it won't overwrite it and your recording
will be erased when the code restarts.