Read and write the autonomous program from disk
This uses the "Drivetrain_Replay_Folder" key in SmartDashboard.  It's
essential that this points at a folder with read and write access on the
robot.  It must be an absolute path, trailing slashes shouldn't matter
however.  The path defaults to /home/lvuser, which isn't _too_ awful but
it will get overwritten by the next recording.

Example:

- Log in as admin@roboRIO-4915-FRC.local
- su lvuser
- mktemp -d
- Copy that path.

This commit also fixes the issue only allowing a recording to be played
once per boot of the roboRIO.